spec -> legacy_spec
This change moves the existing test suite into legacy_spec. The original
test suite had several issues, one of which was the prescence of race
conditions that meant unit tests would occassionally fail for unknown
reasons.

We intend to refactor the two primary classes of Que, Job and Worker,
removing as much of their complexity as possible to make Que easier to
understand. As we do this we'll add tests from legacy_spec back into
spec, ensuring that they are consistent in style and don't introduce
race conditions.

Co-authored-by: Harry Maclean <harrymaclean@gocardless.com>