Refactor Que::Job, Que::Worker
This commit modifies the Job and Worker class to better reflect the
separation of responsibilities between the two concepts. In summary, the
Worker class knows how to acquire and run jobs, while Job knows how to
enqueue, run and destroy itself.

One thing to note is the removal of any parallelism from the Que worker.
We don't use this at GoCardless, and are making the decision to remove
this temporarily. This will be added back in at a later point, but will
be managed externally to the Worker code, as orchestrating workers is
not a responsibility of a worker.

-# Que::Job

Prior to this change, the Job class included a logic about how to lock a
job, which is (in our opinion) essential to the function of a worker. It
also contained code for JSON serialization and Ruby constant lookups,
which are best placed elsewhere.

Now the interface for Que::Job comprises of enqueue, run and destroy.
The assignment of default parameters and the handling of extracting
those attributes from parameters handed to the job is simplified, while
maintaining compatibility with the semantics of the previous Job
implementation.

-# Que::Worker

Previously Worker contained code that tracked a grab-bag of Que module
constants, and responded to class level methods by mutating global state
to create many worker instances. The worker itself did not know how to
lock a job, and defined several (global) class instance variables.

We have now removed all code from Worker that wasn't directly related to
finding and working jobs. The class is now exclusively about working a
job, and repeating that ad infinitum. Instead of using global instance
variables, each Worker is initialized with its configuration, removing
an entire category of errors when tweaking Worker config variables.

Co-authored-by: Harry Maclean <harrymaclean@gocardless.com>