Remove que rake task in favour of binary
This change deprecates the Que rake task (as in chanks upstream) in
favour of a single binary. This is advantageous as we no longer need
maintain two copies of code to configure and kickstart workers.

We remove the --worker-count flag, as we no longer support this. In a
positive spin, we remove a huge number of configuration setters that
interacted with global state that we now encapsulate for each worker.

Co-authored-by: Harry Maclean <harrymaclean@gocardless.com>