Remove helper log method
The Que.log method encourages impedences mismatch between many different
logging interfaces. We don't benefit much from having Que decide for us
what the fields in our logs should be, as we configure these on the
underlying logger anyway.

We can then also remove the log_formatter, as that would be configured
on the logger.