Prevent negative prometheus counter increment
We're not sure of the cause, but for very fast jobs the computation of
duration can sometimes result in a negative elapsed. If passed to
prometheus, then we'll fail our job. We should prevent this at all costs
by adjusting the elapsed to be zero if we detect that it's negative.