Que benchmark
Create a benchmark that can be used to test changes to Que in a
reliable, repeatable fashion. This leverages kubernetes to run several
replicas of sample Que workers, with a script to publish large numbers
of jobs to the que_jobs table with randomised priorities.

We make a small modification to fix an incorrect constant in the bin/que
script. This wasn't noticed before because we always set the parameter
and avoided execution of the bad code path.