Add metrics and improve work loop
We had a performance bug where introducing cursors could cause the Que
worker to trigger the wake_interval sleep when there were jobs remaining
in the queue to be worked. This is because we'd attempt to lock a job
with a job_id > cursor, find no job that matched that condition, and
return to the Que worker that there are no jobs.

This is incorrect, as there are many jobs in the queue, they just didn't
match our initial cursored query. We should instead immediately retry
acquiring a job without the cursor to ensure we don't leave jobs
unworked in the queue.

There was also another bug that was triggered less frequently where we
would return nil when we hit the MVCC race condition. We'd also return
nil here, which would be wrong as it signals there are no jobs left,
when actually we've just raced against another worker. Fixing this
improved throughput by about 10%.

To debug these issues we introduce a variety of new metrics, which
include the following:

  que_worker_sleeping_total
  que_worker_sleeping_seconds_total
  que_worker_job_exists__total
  que_worker_job_exists_seconds_total
  que_worker_job_unlock_total
  que_worker_job_unlock_seconds_total

This allows further breakdown of Que worker behaviour to ensure we cover
the significant factors of Que performance, and can detect when we're
sleeping on the job!