Threaded workers (--worker-count)
Allow creation of a threaded Que worker pool. This used to be in Que,
but was removed during a tidy-up.

The new threaded pool incorporates the work done on stop timeouts to
ensure we mark Que jobs as having failed, with the associated exception,
if the worker is requested to stop but fails to do so in time.

Notable changes include the removal of the --timeout-message flag (given
this can be provided by the Que::JobTimeoutError exception) and the
namespacing of the JobTimeoutError under the Que module. This will
require a small change in QueFailure.