Compact CLI worker options
In `Que::Worker` there are a bunch of class constants
that are used as default values for the kwargs to the
constructor. However, passing nil explicitly as a
kwarg in ruby means that the default is not applied.

```ruby
> def mymeth(foo: :bar); foo; end
=> :mymeth
> mymeth
=> :bar
> mymeth(foo: nil)
=> nil
```

So for users of the que CLI script, omitting options
actually causes failures. This commit fixes it.