Add a timeout to the worker shutdown procedure
This allows us to deal with the situation where a job using the
`NoRetry` strategy takes a long time to complete. Normally this would
result in the process being forcibly killed (SIGKILL) by our scheduler.
By setting this timeout to just under the time between SIGTERM and
SIGKILL, we can ensure that the job fails cleanly, with a clear error
message.