Modify JobTimeoutError's superclass to avoid unguarded rescue blocks.
This works around code that is build to swallow any exception that
descends from StandardError in order (e.g.) to skip bad data to prevent
it from halting an entire batch of work.

Although we should really update that code _not_ to catch exceptions so
generally, that's not going to be easy vs. this change.