Use 'default' for default queue, instead of ''
Que has historically used the empty string '' to represent the default
queue. This causes significant confusion, whereas using the explicit
'default' string would be obvious and clear for anyone who happened to
look.

This commit modifies the default queue value to be 'default'. We do not
include a database migration as this is probably best to do by hand,
updating any jobs that remain in the old '' queue to be 'default' during
the switchover.