Point the Gemfile to the our branch of Prom Client
Ideally we'd do this in the `gemspec`, but you can't do that.
Instead, we remove the version restriction from `gemspec` and add it to
the `Gemfile` here, and in any other clients of `Que`.

This is extremely gross, but unless we properly fork, we have to do this
for now.