Deprioritise worker threads wrt control flow
We're seeing Que jobs that are very CPU intensive cause our liveness
probes to fail, which in turn kills the Que worker. We want to ensure
that our metric endpoints function even when our workers are working
hard, so we deprioritise our worker threads relative to our Que control
loop and metrics server.