Override adapter when custom adapter is specified
The adapter that Que uses to communicate with the database is specified
on the module level. This means that Que cannot be used in projects that
have two databases, as Que can only have one default adapter at a time.
This change allows custome adapters to be specified on the Job level, so
specific jobs can be enqueued and destroyed in the correct database.