Fix default value for cursor_expiry
Add back the default value for cursor_expiry that was removed in #55.

Set it to default to the wake_interval unless provided as an option -
A poll interval of X  suggests you’re ok with not working a job in the
queue for up-to X, which is the effect a cursor would have on which
jobs get worked next.