Assorted fixes for the leaky bucket
The previous commit didn't actually use the observe method of the leaky
bucket. This meant the bucket never tracked any work, so while the
bucket itself worked, it never took any throttling action.

We also provide an improvement to the observe method of the leaky bucket
to ensure we throttle things even when the lock timeout throws an
exception.