Add support for using keyword arguments in job classes
This behaviour works by extracting the keyword arguments to a trailing
hash in the :args array; using a special "kwargs" flag we convert this
trailing hash into a symbol hash that can be double-splatted back onto
the class when calling #run.

Because we're assuming args is a collection, we coerce it to an empty
array here rather than simply `nil` - when we splat it later on
(`*@args`) the behaviour will be the same.