Export metrics for worker groups
Each que "worker" (as an executable) will spawn a worker group, which is
a collection of threads that will end up picking up the jobs on the
configured queue. Practically speaking, each of these threads is a
worker for the configured queue.

We are already exporting metrics for individual worker threads, but we
have no visibility into the worker group itself. In situations where one
or more threads fail in a group, the process is left with less capacity
than originally intended but as it stands we have no visibility into how
often does that happen and when.

To that extent, this commit adds two gauges as Prometheus metrics:

- `que_worker_group_expected_workers_count`: how many workers we wanted
  this process to run
- `que_worker_group_active_workers_count`: how many workers are
  currently active and running