Do not kill the worker thread if we can't checkout a PG connection
It's possible for the adapter to not being able to checkout a new PG
connection. In the case of AR specifically, this can raise a few
different exceptions depending on what's the reason behind the failure.

Currently when AR raise a connection timeout, the thread won't catch
that exception, meaning that it will crash and leave the worker with one
less thread available to process job. In cases where we only use a
single thread (the default case), this means it's possible for a worker
to be up and running but not actually processing any new jobs.

This commit introduces a way to recognise those exceptions and let the
thread retry on failure, so that we're not left with stale workers.