Handle ActiveRecord::StatementInvalid nested exceptions
`ActiveRecord::StatementInvalid` is one of the most generic exceptions AR
can raise and acts as a wrapper for a specific type of errors (usually
::PG::* errors). Unfortunately if AR raises a `StatementInvalid`
exception, the thread will die and the worker will be left with one less
thread.

This commit handles `StatementInvalid` exceptions such that the original
exception that caused it can be handled without letting the thread die.