Remove global fake job w/custom logs
To align with a more modular / composition-based approach to stubbing, I decided to remove the dedicated 'fake job with custom logging'. For example, there is a test needed (adding in the next commit) that needs to both raise an error and add custom logs. It doesn't make sense to add a new class every time