Downgrade que.job_not_found log to debug
This log happens when there is no work on the workers queue and isn't
particularly useful in production. If we want to assess if the queue is
being over provisioned we should instead look at the
SleepingSecondsTotal metric.

I've preserved the postgres_error log at its current level, its likely
this is at the wrong level also and should be higher but its likely we'd
see other more useful errors alongside this.